refactor: rename doc-ref domain to doc-query
Ref. metriport/metriport-internal#699
